Reorganize modules
Static content into static dir
