+++ src/chime_dash/app/static/definitions.py
