+++ src/chime_dash/app/static/footer.py
